makzen-related-products-rs/requirements.txt
flask

pandas

sortedcontainers

mysql.connector

flask_cors

sklearn

sqlalchemy

scipy

scikit-learn==1.0.2